Update Vagrant config to allow testing under Ruby 2.2.0/Linux 3.10
This updates the Vagrant config to allow for testing under Ruby 2.2.0
and Linux kernel 3.10. The old Vagrantfile config is still in place,
testing can be done against an earlier Ruby and kernel version by
manually switching over which VM is used in the Rakefile, since it's
not quite worth automating this yet.

The Vagrantfile was also updated to use the nrel CentOS 6.5 box, since
the Chef provided one doesn't seem to be working with the VirtualBox
guest additions any longer.