Randomize keys to prevent intermittent spec errors
Our `set_timeout` test doesn’t play nicely with the other tests, and,
depending on the order in which the test suite runs, sometimes gums up
the works by expiring the key when no key should be present at all.

In order to get around this issue, we can simply randomize the key names
so each test operates on its own key in isolation from the others.